requirements.txt
pandas
openai
streamlit
faiss-cpu
beautifulsoup4
sentencepiece
sentence-transformers
python-dotenv
langchain
requests
streamlit-chat
html2text
pinecone-client
